+++ test/Json.hs
@@ -1,83 +1,70 @@
 {-# LANGUAGE RankNTypes #-}
 module Json
-  ( jsonTests
-  , jsonPrisms
+  ( jsonPrisms
   ) where
 
-import           Control.Applicative        (liftA2)
-import           Control.Lens               (Prism', preview, review, _Empty)
+import           Control.Applicative  (liftA2)
+import           Control.Lens         (Prism', preview, review, _Empty)
 
 import           Test.Tasty
-import           Test.Tasty.Hedgehog        (testProperty)
-import           Test.Tasty.HUnit           (testCase, (@?=))
+import           Test.Tasty.Hedgehog  (testProperty)
+import           Test.Tasty.HUnit     (testCase, (@?=))
 
-import           Hedgehog                   (Gen, MonadGen, Property, forAll,
-                                             property, (===))
-import qualified Hedgehog.Gen               as Gen
-import qualified Hedgehog.Range             as Range
+import           Hedgehog             (Gen, Property, forAll, property, (===))
+import qualified Hedgehog.Gen         as Gen
+import qualified Hedgehog.Range       as Range
 
-import           Data.HashMap.Lazy          (HashMap)
-import qualified Data.HashMap.Lazy          as HM
+import           Data.Vector          (Vector)
+import qualified Data.Vector          as V
 
-import           Data.Scientific            (Scientific)
-import qualified Data.Scientific            as Sci
-import           Data.Text                  (Text)
+import           Data.HashMap.Strict  (HashMap)
+import qualified Data.HashMap.Strict  as HM
 
-import qualified Data.Attoparsec.ByteString as AB
-import qualified Data.Attoparsec.Text       as AT
+import           Data.Text            (Text)
+
+import qualified Data.Attoparsec.Text as AT
 
 import           Waargonaut.Types
 
-import           Waargonaut                 (_ArrayOf, _Bool, _Number,
-                                             _ObjLazyHashMapOf, _String,
-                                             _TextJson)
+import           Waargonaut           (_ArrayOf, _Bool, _Number, _ObjHashMapOf,
+                                       _String, _TextJson)
 
-import           Types.Json                 (genJson)
+import           Types.Common         (genScientific, genText)
+import           Types.Json           (genJson)
 
-jsonTests :: TestTree
-jsonTests =
-  testGroup "Json types"
-    [ testCase "CommandSepareted's _Empty prism law"
-      $ preview _Empty (review _Empty () :: CommaSeparated WS (JAssoc WS Json)) @?= Just ()
-    , testCase "JObject's _Empty prism law"
-      $ preview _Empty (review _Empty () :: JObject WS Json) @?= Just ()
-    , testCase "JArray's _Empty prism law"
-      $ preview _Empty (review _Empty () :: JArray WS Json) @?= Just ()
-    , testCase "MapLikeObj's _Empty prism law"
-      $ preview _Empty (review _Empty () :: MapLikeObj WS Json) @?= Just ()
-    , testCase "WS's _Empty prism law"
-      $ preview _Empty (review _Empty () :: WS) @?= Just ()
 
-    ]
-
-prismLaw
-  :: ( Eq a
-     , Show a
-     )
-  => Gen a
-  -> Prism' b a
-  -> Property
-prismLaw genA prismA = property $ do
-  a <- forAll genA
+prismLaw :: (Eq a, Show a) => Gen a -> Prism' b a -> Property
+prismLaw genA prismA = property $ forAll genA >>= \a ->
   preview prismA (review prismA a) === Just a
 
-genScientific :: MonadGen m => Maybe Int -> m Scientific
-genScientific lim = either fst fst . Sci.fromRationalRepetend lim
-  <$> Gen.realFrac_ (Range.linearFrac 0.0001 1000.0)
+emptyPrismLaw :: TestTree
+emptyPrismLaw =
+  testGroup "_Empty"
+    [ testCase "CommaSeparated"
+      $ preview _Empty (review _Empty () :: CommaSeparated WS (JAssoc WS Json)) @?= Just ()
+    , testCase "JObject"
+      $ preview _Empty (review _Empty () :: JObject WS Json) @?= Just ()
+    , testCase "JArray"
+      $ preview _Empty (review _Empty () :: JArray WS Json) @?= Just ()
+    , testCase "MapLikeObj"
+      $ preview _Empty (review _Empty () :: MapLikeObj WS Json) @?= Just ()
+    , testCase "WS"
+      $ preview _Empty (review _Empty () :: WS) @?= Just ()
+    ]
 
 objHashMapPrismLaws :: TestTree
-objHashMapPrismLaws = testGroup "Json lazy hashmap prism"
-  [ testProperty "_ObjLazyHashMapOf Scientific"
-    $ prismLaw (genHashMapOf $ genScientific (Just 10)) (_ObjLazyHashMapOf _Number)
+objHashMapPrismLaws = testGroup "Json hashmap prism"
+  [ testProperty "_ObjHashMapOf Scientific"
+    $ prismLaw (genHashMapOf $ genScientific (Just 10)) (_ObjHashMapOf _Number)
 
-  , testProperty "_ObjLazyHashMapOf Text"
-    $ prismLaw (genHashMapOf genUnicodeText) (_ObjLazyHashMapOf _String)
+  , testProperty "_ObjHashMapOf Text"
+    $ prismLaw (genHashMapOf genText) (_ObjHashMapOf _String)
 
-  , testProperty "_ObjLazyHashMapOf Bool"
-    $ prismLaw (genHashMapOf Gen.bool) (_ObjLazyHashMapOf _Bool)
+  , testProperty "_ObjHashMapOf Bool"
+    $ prismLaw (genHashMapOf Gen.bool) (_ObjHashMapOf _Bool)
 
-  , testProperty "_ObjLazyHashMapOf Json"
-    $ prismLaw (genHashMapOf genJson) (_ObjLazyHashMapOf (_String . _TextJson AT.parseOnly))
+  , testProperty "_ObjHashMapOf Json"
+    $ prismLaw (genHashMapOf genJson) (_ObjHashMapOf (_String . _TextJson AT.parseOnly))
   ]
 
 arrayOfPrismLaws :: TestTree
@@ -86,7 +73,7 @@ arrayOfPrismLaws = testGroup "Json array prism"
     $ prismLaw (genListOf $ genScientific (Just 10)) (_ArrayOf _Number)
 
   , testProperty "_ArrayOf Text"
-    $ prismLaw (genListOf genUnicodeText) (_ArrayOf _String)
+    $ prismLaw (genListOf genText) (_ArrayOf _String)
 
   , testProperty "_ArrayOf Bool"
     $ prismLaw (genListOf Gen.bool) (_ArrayOf _Bool)
@@ -97,29 +84,20 @@ arrayOfPrismLaws = testGroup "Json array prism"
 
 jsonPrisms :: TestTree
 jsonPrisms =
-  testGroup "Json 'Prisms'' - OBEY THE LAW (preview p (review p x) == Just x)"
-  [ testProperty "_Bool"
-    $ prismLaw Gen.bool _Bool
-
-  , testProperty "_Number"
-    $ prismLaw (genScientific (Just 10)) _Number
-
-  , testProperty "_TextJson"
-    $ prismLaw genJson (_TextJson AT.parseOnly)
-
-  , testProperty "_String"
-    $ prismLaw genUnicodeText _String
+  testGroup "Json 'Prisms'' must OBEY THE LAW: (preview p (review p x) == Just x)"
+  [ testProperty "_Bool"     $ prismLaw Gen.bool _Bool
+  , testProperty "_Number"   $ prismLaw (genScientific (Just 10)) _Number
+  , testProperty "_TextJson" $ prismLaw genJson (_TextJson AT.parseOnly)
+  , testProperty "_String"   $ prismLaw genText _String
 
   , arrayOfPrismLaws
   , objHashMapPrismLaws
+  , emptyPrismLaw
   ]
 
 genHashMapOf :: Gen a -> Gen (HashMap Text a)
-genHashMapOf genElem = HM.fromList <$> genListOf tup
-  where tup = liftA2 (,) genUnicodeText genElem
+genHashMapOf genElem = HM.fromList <$> Gen.list (Range.linear 0 100) tup
+  where tup = liftA2 (,) genText genElem
 
-genListOf :: Gen a -> Gen [a]
-genListOf genElem = Gen.list (Range.linear 0 100) genElem
-
-genUnicodeText :: Gen Text
-genUnicodeText = Gen.text (Range.linear 0 100) Gen.unicode
+genListOf :: Gen a -> Gen (Vector a)
+genListOf genElem = V.fromList <$> Gen.list (Range.linear 0 100) genElem